Refactor: Allow constructing Fetcher directly from SubrequestChannel.
I found myself writing multiple implementations identical to `WorkerStubEntrypointOutgoingFactory`. Let's just bake it in.

Note: The careful observer may notice that this change no longer sets the `isInHouse` flag dynamically-loaded worker requests (by omitting it, thus accepting the default `false`, whereas previously it was explicitly set `true`). This is intentional, as I remembered what "in house" is meant to mean, and realized it ought to be false here. The `inHouse` flag is meant to apply to subrequests where the application isn't really aware it's making a subrequest, such as Workers KV requests. It is NOT normally applied to service binding requests, which are entirely visible to the app. Requests to dynamically-loaded workers seem like they should behave like service bindings.