Add readable-source/writable-sink adapters
This is the first step towards a significant refactoring of the
streams implementation with the goal of making it safer and more
performance.

This commits adds four new classes:

* `ReadableStreamSourceKjAdapter`: Adapts a ReadableStream to the
  `ReadableStreamSource` API, fully encapsulating the JS interactions
  behind the KJ-friendly interface.
* `ReadableStreamSourceJsAdapter`: Adapts a `ReadableStreamSource`
  to a more friendly JS interface.
* `WritableStreamSinkKjAdapter`: Adapts a WritableStream to the
  `WritableStreamSink` API, fully encapsulating the JS interactions
  behind the KJ-friendly interface.
* `WritableStreamSinkJsAdapter`: Adapts a `WritableStreamSink`
  to a more friendly JS interface.

Also adds tests.

This commit does not update any existing code to use these adapters
yet, that will come in follow-up commits.

There are several key goals for this refactoring:

* Make the streams implementation safer by using the new checked
  Queue for buffering operations and removing direct reliance on
  raw references.
* Reducing the overall complexity of the streams implementation
  through better encapsulation and separation of concerns.

These require very careful review to ensure the behavior is correct
before we start switching existing code over to use these adapters.
In future commits I will be incrementally removing the old code
and switching to use these adapters.